EClassiifier will now be used to define type
- Previously the instnace containing the data needed to be traversed
with ocl and result in an actual object to find the type, now only the
type from the ocl statement is used. This means that a statement that
might not produce a value because the index in the list is out of range
will still allow me to get the inteded type from the statement.